move classes to new package
