+++ app/src/main/java/com/mensinator/app/business/notifications/IAndroidNotificationScheduler.kt
@@ -1,4 +1,4 @@
-package com.mensinator.app.business
+package com.mensinator.app.business.notifications
 
 import java.time.LocalDate
