+++ app/src/test/java/com/mensinator/app/business/NotificationSchedulerTest.kt
@@ -1,6 +1,8 @@
 package com.mensinator.app.business
 
 import android.content.Context
+import com.mensinator.app.business.notifications.IAndroidNotificationScheduler
+import com.mensinator.app.business.notifications.NotificationScheduler
 import com.mensinator.app.data.Setting
 import com.mensinator.app.settings.IntSetting
 import com.mensinator.app.settings.StringSetting
