+++ gradle/libs.versions.toml
@@ -17,7 +17,7 @@ kizitonwoseCalendar = "2.6.2"
 mockk = "1.13.17"
 material3 = "1.3.1"
 material3-adaptive = "1.1.0"
-coroutines = "1.3.9"
+coroutines = "1.10.1"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "core" }
